requirements.txt
asgiref==3.7.2
Django==4.2.2
django-ckeditor==6.6.1
django-js-asset==2.1.0
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
